kolla_k8s/requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6 # Apache-2.0
Babel>=1.3 # BSD
cliff!=1.16.0,!=1.17.0,>=1.15.0 # Apache-2.0
dcos>=0.1.3 # Apache-2.0
docker-py>=1.6.0 # Apache-2.0
Jinja2>=2.8 # BSD License (3 clause)
kazoo>=2.2 # Apache-2.0
netifaces>=0.10.4 # MIT
oslo.config>=3.7.0 # Apache-2.0
oslo.utils>=3.5.0 # Apache-2.0
oslo.log>=1.14.0 # Apache-2.0
PrettyTable<0.8,>=0.7 # BSD
PyYAML>=3.1.0 # MIT
retrying!=1.3.0,>=1.2.3 # Apache-2.0
six>=1.9.0 # MIT
